PIM-7263: reworked service and command
