+++ CHANGELOG-2.2.md
@@ -1,3 +1,9 @@
+# 2.2.x
+
+## BC Breaks
+
+- Rename command `pim:catalog:remove-wrong-values-on-variant-products` to `pim:catalog:remove-wrong-boolean-values-on-variant-products` (but an alias is still here so calling it from the old name will still work)
+
 # 2.2.2 (2018-03-29)
 
 ## BC Breaks
