+++ src/Pim/Bundle/CatalogBundle/Resources/config/updaters.yml
@@ -49,6 +49,8 @@ parameters:
     pim_catalog.updater.remover.multiselect_value.class:      Pim\Component\Catalog\Updater\Remover\MultiSelectAttributeRemover
     pim_catalog.updater.category_translatable.class:          Pim\Component\Catalog\Localization\CategoryUpdater
 
+    pim_catalog.command.cleaner.wrong_boolean_value_on_variant_product.class: Pim\Bundle\CatalogBundle\Command\Cleaner\WrongBooleanValuesOnVariantProductCleaner
+
 services:
     # Updaters
     pim_catalog.updater.product:
@@ -179,6 +181,9 @@ services:
             - '@pim_catalog.repository.group_type'
             - '@pim_catalog.repository.attribute'
             - '@pim_catalog.query.product_query_builder_factory'
+
+    pim_catalog.command.cleaner.wrong_boolean_value_on_variant_product:
+        class: '%pim_catalog.command.cleaner.wrong_boolean_value_on_variant_product.class%'
 
     # Setters
     pim_catalog.updater.setter.registry:
